refactor and move deprecated Collection components
